Creates certs during build.
